requirements.txt
flask
flask-migrate
flask-sqlalchemy
pymysql
